requirements/constraints.txt
bson
cerberus
django>=1.7
djangorestframework>=3.5
marshmallow
pyschemes
PyYAML
six
tablib
wtforms

